weather_forecast_project/requirements.txt
requests
pandas
seaborn
matplotlib
python-dotenv